Extract CallButton component from nav bar & implement call button for mobile
